Split travis build between install deps and run tests
This makes our builds fail as soon as we have any problems installing the
dependencies, instead of going ahead and running the tests. This saves Travis'
resources and our time.

The only problem I've got is that I was unable to run Jetty-Solr's
configuration on travis-install-dependencies. Probably something to do with
Bash's "set -e", but I was unable to pinpoint exactly, so I left it into
travis-run-tests. No harm done.